refactor : controller 로직을 main으로 이동
